+++ modules/components/containers/RuleContainer.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import {getFieldConfig} from "../../utils/configUtils";
 import {pureShouldComponentUpdate} from "../../utils/renderUtils";
 import {connect} from 'react-redux';
+const classNames = require('classnames');
 
 
 export default (Rule) => {
@@ -86,9 +87,13 @@ export default (Rule) => {
       const fieldConfig = getFieldConfig(this.props.field, this.props.config);
       const _isGroup = fieldConfig && fieldConfig.type == '!struct';
 
+      const valueError = this.props.valueError;
+      const oneValueError = valueError && valueError.toArray().filter(e => !!e).shift() || null;
+      const hasError = oneValueError != null;
+
       return (
         <div
-          className={'group-or-rule-container rule-container'}
+          className={classNames('group-or-rule-container', 'rule-container', hasError ? 'rule-with-error' : null)}
           data-id={this.props.id}
         >
         {[
@@ -107,7 +112,7 @@ export default (Rule) => {
             selectedOperator={this.props.operator || null}
             value={this.props.value || null}
             valueSrc={this.props.valueSrc || null}
-            valueError={this.props.valueError}
+            valueError={this.props.valueError || null}
             operatorOptions={this.props.operatorOptions}
             config={this.props.config}
             treeNodesCnt={this.props.treeNodesCnt}
@@ -129,7 +134,7 @@ export default (Rule) => {
             selectedOperator={this.props.operator || null}
             value={this.props.value || null}
             valueSrc={this.props.valueSrc || null}
-            valueError={this.props.valueError}
+            valueError={this.props.valueError || null}
             operatorOptions={this.props.operatorOptions}
             config={this.props.config}
             treeNodesCnt={this.props.treeNodesCnt}