+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - Added displaying of rule validation errors, see `showErrorMessage` in config.settings
   - Added `QbUtils.isValidTree()`
   - `validateValue` moved from widget settings to field's `fieldSettings`
+  - Added ESLint (unused vars & props are off for now)
 - 2.0.11
   - Added `starts_with`, `ends_with` operators for text type
 - 2.0.10